fix(SessionTokenAuthenticator): ensure kwargs are passed
kwargs are not currently passed to `_create_component_from_model` when
creating authenticators used by the selective authenticator. This
results in a TypeError when creating a session token authenticator since
the required  `name` argument passed down from the declarative
stream is not avaliable.

This commits ensures that any kwargs are passed into the
`_create_component_from_model` method when instantiating a
SelectiveAuthenticator.